+++ app/Database/Migrations/2026-08-19-235000_AddCodigoRastreioToPedidos.php
@@ -8,19 +8,23 @@ class AddCodigoRastreioToPedidos extends Migration
 {
     public function up(): void
     {
-        $this->forge->addColumn('pedidos', [
-            'codigo_rastreio' => [
-                'type'       => 'VARCHAR',
-                'constraint' => 50,
-                'null'       => true,
-                'default'    => null,
-                'after'      => 'status_pagamento',
-            ],
-        ]);
+        if (!$this->db->fieldExists('codigo_rastreio', 'pedidos')) {
+            $this->forge->addColumn('pedidos', [
+                'codigo_rastreio' => [
+                    'type'       => 'VARCHAR',
+                    'constraint' => 50,
+                    'null'       => true,
+                    'default'    => null,
+                    'after'      => 'status_pagamento',
+                ],
+            ]);
+        }
     }
 
     public function down(): void
     {
-        $this->forge->dropColumn('pedidos', 'codigo_rastreio');
+        if ($this->db->fieldExists('codigo_rastreio', 'pedidos')) {
+            $this->forge->dropColumn('pedidos', 'codigo_rastreio');
+        }
     }
 }
